pyproject.toml
[tool.poetry]
name = "api_books"
version = "0.1.0"
description = ""
authors = ["Henrique Andrade <henrique.ccda@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"

[tool.poetry.dev-dependencies]
pytest-django = "^4.5.2"
flake8 = "^4.0.1"
ipython = "^8.4.0"
pytest-mock = "^3.8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
